Fixed some bugs preventing compilation on OS X and Linux.  Working on dealing with rapid state transitions (mainly caused by the stupid character bounce).  ModelInstance now defaults to the new animation system if a controller is present, otherwise it falls back to the old system.
Former-commit-id: f3951c6e8c96f89e05e235f8af5feb5c0a835dc8

+++ Sample/source/ActorAids.cpp
@@ -69,7 +69,7 @@ bool ActorAids::Update() {
 	//right now this done stupidly
 	//by randomly selecting an action
 	auto randomAction = actionList.begin();
-	unsigned int randChoice = Utilities::random(0u,actionList.size());
+    unsigned int randChoice = Utilities::random((unsigned int) 0, (unsigned int) actionList.size());
 	if (randChoice >= actionList.size())
 		randChoice = actionList.size()-1;
 	advance(randomAction,randChoice);
